Fix bearer auth guard for requests without body

+++ src/modules/comments/router/handlers/deleteCommentHandler.ts
@@ -5,6 +5,14 @@ import {HTTP_STATUSES} from "../../../../shared/constants/httpStatuses";
 import {commentsService} from "../../application/commentsService";
 
 export const deleteCommentHandler = async (req: RequestWithPathParamsModel<IdPathParamsModel>, res: Response<void>) => {
+    const {userId} = req;
+
+    if (!userId) {
+        res.sendStatus(HTTP_STATUSES.UNAUTHORIZED);
+
+        return;
+    }
+
     const isDeleted = await commentsService.delete(req.params.id);
 
     if (!isDeleted) {
